requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy==2.3.0
scikit-learn>=1.4.0
shap>=0.45.0
plotly>=5.20.0
matplotlib>=3.8.0
google-generativeai>=0.5.0
reportlab>=4.1.0
joblib>=1.3.0
requests>=2.31.0